Problem: Total power calculation are lost
Solution: normalize the output from NUT

UPSes provides different list of values. We are completing the list
with similar values with the idea that approximate value is better
than nothing.

For example if UPS does the measurement on input only (input.L1.realpower),
we produce also output with the same value (output.L1.realpower).

Signed-off-by: Michal Vyskocil <michalvyskocil@eaton.com>