Reduce header file dependencies
It is hard to work with the repository when changing one header file
results in a full rebuild. Theare are two causes for this: The
generated fty_nut_classes.h file, which is completely useless, and the
fty_nut_library.h file (also generated), which besides pulling in the
"public" headers, also defines the FTY_NUT_EXPORT macro. However, we are
building the code with default visibility, so FTY_NUT_EXPORT is a no-op
and can be safely removed. If we ever decide to use -fvisibility-hidden,
an option may be to make all the classes private, because this
repository's deliverable is the two services and not a shared library.