Add a class for managing states of the asset list
The current model with the state file is flawed in two ways: After the
agent startup, we simply read the state file and assume that it is up to
date, which is not the case if new assets appear while the agent is
being restarted. Another problem is that this requires all the threads
listen on the ASSETS stream and apply essentially the same changes to
their copy of the asset list. The idea behind the new StateManager class
is that there will be a single thread that populates the StateManager
with existing assets at startup and keeps updating it as per messages
received on the ASSETS stream and the other threads will simply retrieve
the current state from the StateManager.