Simplify malamute connection setup
Pass the endpoint name as an argument to the actor function and move the
stream setup logic to the actors themselves. This not only reduces code
size, but also makes sure that we enter the main loop with a workable
malamute connection. While at it, fix a few memory leaks in error paths.