Prepare the AssetState API for licensing limitation support
Add two new methods, getAllPowerDevices() and getAllSensors() to return
a map of all known power / sensor assets. The existing methods will be
changed to only return the assets allowed by the current licence. The
fty-nut-configurator agent needs this, in order to maintain
configuration of all assets even if not all of them are currently
allowed. The other agents will only work with devices allowed by the
licence.