fty-nut-configurator: Run systemctl actions in batches
In the main loop, process incoming messages as long as there are any and
call Autoconfig::onPoll() only if there is nothing left in the queue.
In NUTConfigurator::configure() and NUTConfigurator::erase(), only
update the configuration snippets and do the restart the services after
all changes in the current round have been processed. This provides a
considerable speed up when importing large number of devices.